requirements.txt
requests==2.18.4
python-telegram-bot==10.1.0
telethon==1.0.1